Detect, install and use gb
Use a specific gb version, pull it from github (outage there makes this
not work, consider precompiling GB for every supported version of Go).

@davecheney indicates that gb should be compiled with the same version
of Go that is in use (at least for now), so we don't bother having a
version of gb available until we've determined that.

In lieu of having gb available for detection we look for a `src/`
directory containing any `*.go` files and if we find that then we assume
it's a gb project.

Cache gb and if the gb version changes we nuke any previously cached
versions.

Like Godep, we also support `-tags heroku` and linker symbols.

This also changes the order of detection, moving .godir and
$build/Godeps (file) detection to after supported methods to bring
compile and detect more inline with each other.

Make detect a little more discriminating as well, previously we passed
for any project with a *.go file in it, which could be bad at some point.